Moved MinMaxInt to root of scripts, made MinMaxFloat
